fix: entity dto base

+++ nupkg/pack.ps1
@@ -19,8 +19,10 @@ foreach($project in $projects) {
     ###########################################################################################################
 
     # Clean previous build and compile newly
-    Remove-Item -Recurse (Join-Path $projectFolder "bin/Release")
-    & dotnet msbuild -property:Configuration=Release -property:SourceLinkCreate=true
+    if(-Not(Test-Path -Path (Join-Path $projectFolder "bin/Release"))){
+        Remove-Item -Recurse (Join-Path $projectFolder "bin/Release")
+    }
+    dotnet msbuild -property:Configuration=Release -property:SourceLinkCreate=false
     & dotnet pack -c Release --no-build --include-symbols
     
     ###########################################################################################################
